Stepping now uses function(s), log of recent tps#s
Stepping now uses step_ope() or step_close() function
Added ability to see if last x iterations of stepping are the same and if last x iterations of tps values are the same. This was done in hopes to use for ensuring throttle moves if stepping. Need to find a reliable way of preventing stepper from moving if throttle is not moving for last x iterations while not breaking program.
Various minor changes.